chore: move tests to correct namespace
